+++ tests/models/identity/config.pbtxt
@@ -11,7 +11,6 @@
 
 name: "identity"
 backend: "python"
-
 max_batch_size: 1
 
 input [
@@ -19,7 +18,6 @@ input [
     name: "INPUT0"
     data_type: TYPE_FP32
     dims: [ 3, 256, 256 ]
-
   }
 ]
