+++ tests/regression/test_fake.py
@@ -9,22 +9,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import unittest
 from typing import Dict
 
 import cupy
-import torch
 from monai.transforms import Compose, Identityd
 
 from monaistream.compose import StreamCompose
 from monaistream.filters import FilterProperties, NVVideoConvert, TransformChainComponent, TransformChainComponentCupy
-from monaistream.filters.infer import NVInferServer
 from monaistream.sinks import FakeSink
 from monaistream.sources import FakeSource
 
 
-class TestRBGAWithFake(unittest.TestCase):
+class TestWithFake(unittest.TestCase):
     def test_shortcircuit(self):
         pipeline = StreamCompose(
             [
@@ -115,43 +112,3 @@ class TestRBGAWithFake(unittest.TestCase):
             ]
         )
         pipeline()
-
-    def test_customuserdata(self):
-        def assert_copy_equal(inputs: Dict[str, torch.Tensor]):
-            self.assertEqual(inputs["ORIGINAL_IMAGE"], inputs["OUTPUT0"])
-            return inputs
-
-        infer_server_config = NVInferServer.generate_default_config()
-        infer_server_config.infer_config.backend.trt_is.model_repo.root = os.path.join(os.getcwd(), "tests", "models")
-        infer_server_config.infer_config.backend.trt_is.model_name = "identity"
-        infer_server_config.infer_config.backend.trt_is.version = "1"
-        infer_server_config.infer_config.backend.trt_is.model_repo.log_level = 0
-
-        chain = StreamCompose(
-            [
-                FakeSource(),
-                NVVideoConvert(
-                    FilterProperties(
-                        format="RGBA",
-                        width=256,
-                        height=256,
-                        channels=4,
-                        framerate=(32, 1),
-                    )
-                ),
-                NVInferServer(
-                    config=infer_server_config,
-                ),
-                TransformChainComponent(
-                    output_label="ORIGINAL_IMAGE",
-                    transform_chain=assert_copy_equal,
-                ),
-                FakeSink(),
-            ]
-        )
-        chain()
-
-
-if __name__ == "__main__":
-    t = TestRBGAWithFake()
-    t.test_customuserdata()
